requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
Pillow==10.1.0
chromadb==0.4.18
python-dotenv==1.0.0
groq>=0.31.0
numpy==1.26.4
httpx>=0.24.1,<0.25.0
poppler-utils
pydantic==2.4.2
pydantic-settings==2.0.3
starlette==0.27.0
aiofiles==23.2.1
tqdm==4.66.1
colorama==0.4.6
sentence-transformers
